refactor(ui): split ui.rs into ui/ module, extract theme submodule
Renames src/ui.rs → src/ui/mod.rs and lifts the inline `mod theme`
block (~200 lines) into src/ui/theme.rs. Pure file reorganization;
no behavior change. Snapshots move to src/ui/snapshots/ to track
their source file.

First step toward breaking up the 5,100-line ui.rs along the
boundaries already implicit in the file: chrome widgets, per-tab
renderers, state types, formatters. Each subsequent module gets
its own follow-up so each diff stays reviewable.
